core/vm: move native asset contracts to separate file
